hardcoding begins

+++ app/src/main/java/Modules/Route.java
@@ -14,7 +14,7 @@ public class Route {
     public LatLng endLocation;
     public String startAddress;
     public LatLng startLocation;
-    public int rating;
+    public double rating;
     public String jsonRAW;
     public List<LatLng> points;
     public List<String> placeIds;
